fix: GetTeamResponseDTO 생성자 인자 값 및 isNew까지 모두 해결하도록 수정

+++ src/main/java/com/kuit/conet/jpa/repository/TeamMemberRepository.java
@@ -32,8 +32,8 @@ public class TeamMemberRepository {
     }
 
 
-    public Long getCount(Long id) {
-        return em.createQuery("select count(tm) from TeamMember tm where tm.team.id=:id", Long.class)
+    public Integer getCount(Long id) {
+        return em.createQuery("select count(tm) from TeamMember tm where tm.team.id=:id", Integer.class)
                 .setParameter("teamId", id)
                 .getSingleResult();
     }
